Fix sokol_app.h not building on macOS 10.12
NSPasteboardTypeFileURL is not available on macOS 10.12. This commit
wraps the two usages of it in a preprocessor guard. This will cause
file drag and drop operations to not work on 10.12, but at least it will
build. A way to implement similar drag and drop on 10.12 and earlier
exists, but this commit doesn't implement it. I don't know exactly what
it would be, and I wouldn't know how to test it with sokol_app.h