sokol_app.h iOS: low- vs high-dpi fixes for Metal and GL.
MTKView.contentScaleFactor seems to be ignored now, which means
that rendering on iOS was always in HighDPI. Instead of relying
on MTKView contentScaleFactor, the drawableSize is now set
explicitly.

When high_dpi is enabled, the dpi_scale factor is now queried
from iOS (via UIScreen.mainScreen.nativeScale).

MSAA is now supported for the iOS GL backend.