sokol_gfx.h: change update-image validation size check
It is now only checked that the provided data blob has at least
the required data size, but can be bigger. This is necessary
because backends have different surface data alignment
rules for (PVRTC) compressed textures. This change requires
further fixes in other backends where the provided surface
size is used for copying operations instead of the actual
surface size!