sokol_app.h gl: make GL context current in NSOpenGLView.reshape
Not sure if this is a regression in macOS, or never worked:

The SAPP_EVENTTYPE_RESIZED event may be fired from within the
NSOpenGLView.reshape method, which doesn't seem to have a GL
context bound, so calling any GL function from within
the RESIZED handler fails. The commit simply makes the
NSOpenGLView's GL context current in the reshape function.