[sokol_app] proxy javascript function to the main thread when needed
## What does this PR do
This commit changes the javascript defined by sokol to allow proxying them when necessary and enable the use of the proxy_to_pthread option.
Specificallt, it replaces a few call to the `EM_JS` macro to the `EM_JS_MAIN_THREAD`

## What problem does it solve?
When compiled with `proxy_to_pthread` ([generally recommended](https://emscripten.org/docs/porting/pthreads.html#blocking-on-the-main-browser-thread) by emscripten documentation), the "main" app thread is running separately from the browser event loop. This guarantee the browser won't become unresponsive because of a deadlock.
This option is particularly convenient when porting code that might block for any event on the main thread.

With this option, the main app thread is not able to access a few javascript features such as the `window` object. The solution for that is called [proxying]()https://emscripten.org/docs/porting/pthreads.html#proxying), where javascript function calls are move to the main thread and proxied on the worker thread.

# Potential caveat
Since there is no utility `EM_JS_MAIN_THREAD` in emscripten (that would define a javascript function callable from c++ proxied to main), I had to define the corresponding function in C and make them call inline asm via the `MAIN_THREAD_EM_ASM`. This means that the generated function in the javascript side won't be named with the name defined in C, instead they will have random name.

Hence, if those function were intended to be called from the javascript side, this PR would be a problem. I don't see any call like this so far, but the fact that they were not profixed with an underscore as other internal function make me doubt.