sokol_gfx.h gl: implement workaround for Chrome v127 regression (fixes #1085)
Works around a Chrome 127 and Safari Tech Preview regression on macOS
which broke sokol-gfx offscreen rendering. The bug is triggered
by setting the MAX_TEXTURE_LEVEL, but not calling glTexStorage
to allocate storage for immutable textures.

The workaround calls glTexStorage, but currently only on Emscripten.

A proper fix would be to rewrite texture initialization to generally
use glTexStorage + glTexSubImage, but this requires a fallback for
macOS because GL 4.1 doesn't have the glTexStorage call yet.