sokol_gl.h: set default write_mask to RGBA
Setting the write_mask from RGB to RGBA allows for using sokol_gl to draw to a render target (when creating a pipeline with make_context_pipeline), and retain alpha values. Without this change,  render textures written to with sokol_gl will truncate the alpha information, and will appear invisible when blending is enabled.